requirements.txt
# SMC library (existing)
pandas>=2.0.2
numpy>=1.24.3
numba>=0.58.1

# Bot - core
requests>=2.31
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
aiosqlite>=0.20

# Bot - UI
nicegui>=2.0
plotly>=5.18

# Bot - Data Provider (Option A: MT5)
# MetaTrader5>=5.0.45  # Windows only, uncomment if using MT5

# Bot - Data Provider (Option B: OANDA)
# (uses requests, already included above)
